app/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
dbfread>=2.0.7
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6
pandas
